+++ src/navigation/forgottenPassword/ForgottenPasswordModalNavigator.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent } from 'react'
 import { createStackNavigator } from '@react-navigation/stack'
 import { headerBlank } from '../../styles/navigationAppearance'
 import { ForgottenPasswordModalNavigatorParamList } from './ForgottenPasswordModalNavigatorParamList'
-import ForgottenPasswordScreen from '../../screens/authentication/ForgottenPasswordScreen'
+import ForgottenPasswordScreen from '../../screens/forgottenPassword/ForgottenPasswordScreen'
 
 const Stack = createStackNavigator<ForgottenPasswordModalNavigatorParamList>()
