+++ src/navigation/onboarding/OnboardingNavigator.tsx
@@ -1,10 +1,10 @@
 import { createStackNavigator } from '@react-navigation/stack'
 import React from 'react'
 import { FunctionComponent } from 'react'
-import LoginScreen from '../../screens/authentication/LoginScreen'
+import LoginScreen from '../../screens/login/LoginScreen'
 import OnboardingScreen from '../../screens/onboarding/OnboardingScreen'
 import { headerBlank } from '../../styles/navigationAppearance'
-import SignUpScreen from '../../screens/authentication/SignUpScreen'
+import SignUpScreen from '../../screens/signUp/SignUpScreen'
 import { OnboardingNavigatorParamList } from './OnboardingNavigatorParamList'
 
 const Stack = createStackNavigator<OnboardingNavigatorParamList>()
